+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Added
 ~~~~~
 
 - Stateful testing via Open API links for the ``pytest`` runner. `#616`_
+- Support for stateful testing, using only explicit examples from the schema. `#638`_
 
 Changed
 ~~~~~~~
@@ -1369,6 +1370,7 @@ Fixed
 .. _#647: https://github.com/schemathesis/schemathesis/issues/647
 .. _#641: https://github.com/schemathesis/schemathesis/issues/641
 .. _#640: https://github.com/schemathesis/schemathesis/issues/640
+.. _#638: https://github.com/schemathesis/schemathesis/issues/638
 .. _#631: https://github.com/schemathesis/schemathesis/issues/631
 .. _#629: https://github.com/schemathesis/schemathesis/issues/629
 .. _#622: https://github.com/schemathesis/schemathesis/issues/622